postblock2/q5/requirements.txt
google-cloud-storage
pandas
transformers
torch
scikit-learn
emoji
sentencepiece
seaborn
matplotlib
